requirements.txt
# Список зависимостей для проекта
requests==2.28.1          # Библиотека для HTTP-запросов
pytest==7.2.0             # Фреймворк для тестирования
allure-pytest==2.12.0     # Интеграция Allure с pytest
pytest-mock==3.11.1       # Моки для pytest